+++ AnvilCommon/Source/Utils/Util.hpp
@@ -15,9 +15,9 @@ namespace AnvilCommon::Utils
 	public:
 		static bool HasCommandLine(std::string p_Command);
 
-		static bool PatchAddressInMemory(uint64_t p_Offset, std::string p_HexString, int32_t p_Length);
+		static bool PatchAddressInMemory(uint32_t p_Offset, std::string p_HexString, int32_t p_Length);
 		static bool PatchAddressInMemory(void* p_Address, std::string p_HexString, int32_t p_Length);
-		static bool PatchAddressInFile(uint64_t p_Offset, std::string p_HexString, int32_t p_Length);
+		static bool PatchAddressInFile(uint32_t p_Offset, std::string p_HexString, int32_t p_Length);
 		static bool PatchAddressInFile(void* p_Offset, std::string p_HexString, int32_t p_Length);
 		
 		static void WriteCall(void *p_Address, void *p_NewFunction);
